test: create system tests
