+++ .circleci/config.yml
@@ -14,6 +14,10 @@ install_and_test: &install_and_test
         name: Test
         command: npm test
         when: always
+    - run:
+        name: System Tests
+        command: npm run system-tests
+        when: always
         environment:
           GOOGLE_APPLICATION_CREDENTIALS: /root/project/.circleci/key.json
     - run:
