+++ package.json
@@ -16,7 +16,7 @@
     "fix": "gts fix",
     "prepare": "npm run compile",
     "pretest": "npm run compile",
-    "posttest": "npm run check"
+    "system-tests": "mocha build/test/system-tests"
   },
   "keywords": [
     "google",
