+++ test/system-tests/kitchen.ts
@@ -1,0 +1,72 @@
+import * as assert from 'assert';
+import * as fs from 'fs';
+import * as r from 'request';
+const upload = require('../../src');
+
+it('should work', (done) => {
+  let uploadSucceeded = false;
+  fs.createReadStream('daw.jpg')
+    .on('error', done)
+    .pipe(upload({
+      bucket: 'stephen-has-a-new-bucket',
+      file: 'daw.jpg',
+      metadata: {
+        contentType: 'image/jpg'
+      }
+    }))
+    .on('error', done)
+    .on('response', function (resp) {
+      uploadSucceeded = resp.statusCode === 200;
+    })
+    .on('finish', function () {
+      assert.strictEqual(uploadSucceeded, true);
+      done();
+    });
+});
+
+it('should resume an interrupted upload', function (done) {
+  fs.stat('daw.jpg', function (err, fd) {
+    assert.ifError(err);
+
+    const size = fd.size;
+
+    const doUpload = function (opts, callback) {
+      let sizeStreamed = 0;
+      let destroyed = false;
+
+      const ws = upload({
+        bucket: 'stephen-has-a-new-bucket',
+        file: 'daw.jpg',
+        metadata: {
+          contentType: 'image/jpg'
+        }
+      });
+
+      fs.createReadStream('daw.jpg')
+        .on('error', callback)
+        .on('data', function (chunk) {
+          sizeStreamed += chunk.length;
+
+          if (!destroyed && opts.interrupt && sizeStreamed >= size / 2) {
+            // stop sending data half way through
+            destroyed = true;
+            this.destroy();
+            ws.destroy(new Error('Interrupted'));
+          }
+        })
+        .pipe(ws)
+        .on('error', callback)
+        .on('metadata', callback.bind(null, null));
+    };
+
+    doUpload({ interrupt: true }, function (err) {
+      assert.strictEqual(err.message, 'Interrupted');
+
+      doUpload({ interrupt: false }, function (err, metadata) {
+        assert.ifError(err);
+        assert.equal(metadata.size, size);
+        done();
+      });
+    });
+  });
+});
