+++ test/test.ts
@@ -68,83 +68,6 @@ describe('gcs-resumable-upload', function () {
     mockery.disable();
   });
 
-  it('should work', function (done) {
-    this.timeout(90000);
-
-    let uploadSucceeded = false;
-
-    requestMock = _request;
-
-    fs.createReadStream('daw.jpg')
-      .on('error', done)
-      .pipe(upload({
-        bucket: 'stephen-has-a-new-bucket',
-        file: 'daw.jpg',
-        metadata: {
-          contentType: 'image/jpg'
-        }
-      }))
-      .on('error', done)
-      .on('response', function (resp) {
-        uploadSucceeded = resp.statusCode === 200;
-      })
-      .on('finish', function () {
-        assert.strictEqual(uploadSucceeded, true);
-        done();
-      });
-  });
-
-  it('should resume an interrupted upload', function (done) {
-    this.timeout(90000);
-
-    requestMock = _request;
-
-    fs.stat('daw.jpg', function (err, fd) {
-      assert.ifError(err);
-
-      const size = fd.size;
-
-      const doUpload = function (opts, callback) {
-        let sizeStreamed = 0;
-        let destroyed = false;
-
-        const ws = upload({
-          bucket: 'stephen-has-a-new-bucket',
-          file: 'daw.jpg',
-          metadata: {
-            contentType: 'image/jpg'
-          }
-        });
-
-        fs.createReadStream('daw.jpg')
-          .on('error', callback)
-          .on('data', function (chunk) {
-            sizeStreamed += chunk.length;
-
-            if (!destroyed && opts.interrupt && sizeStreamed >= size / 2) {
-              // stop sending data half way through
-              destroyed = true;
-              this.destroy();
-              ws.destroy(new Error('Interrupted'));
-            }
-          })
-          .pipe(ws)
-          .on('error', callback)
-          .on('metadata', callback.bind(null, null));
-      };
-
-      doUpload({ interrupt: true }, function (err) {
-        assert.strictEqual(err.message, 'Interrupted');
-
-        doUpload({ interrupt: false }, function (err, metadata) {
-          assert.ifError(err);
-          assert.equal(metadata.size, size);
-          done();
-        });
-      });
-    });
-  });
-
   it('should just make an upload URI', function (done) {
     this.timeout(10000);
